+++ sdk/storage/azblob/CHANGELOG.md
@@ -1,15 +1,17 @@
 # Release History
 
-## 1.3.3 (Unreleased)
+## 1.4.0-beta.1 (Unreleased)
 
 ### Features Added
+* Updated service version to `2024-05-04`.
 
 ### Breaking Changes
 
 ### Bugs Fixed
 
 ### Other Changes
-* Updated `azidentity` version to `1.5.2`
+* Updated `azidentity` version to `1.6.0`
+* Updated `azcore` version to `1.12.0`
 
 ## 1.3.2 (2024-04-09)