+++ sdk/storage/azdatalake/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History
 
+## 1.2.0-beta.1 (Unreleased)
+
+### Features Added
+* Added ACL response headers in GetBlobProperties API for Files.
+* Updated service version to `2024-05-04`
+
+### Other Changes
+* Updated `azidentity` version to `1.6.0`
+* Updated `azcore` version to `1.12.0`
+
 ## 1.1.3 (2024-06-13)
 
 ### Bugs Fixed