+++ sdk/storage/azfile/CHANGELOG.md
@@ -1,15 +1,18 @@
 # Release History
 
-## 1.2.3 (Unreleased)
+## 1.3.0-beta.1 (Unreleased)
 
 ### Features Added
+* Add Rename Support to List Ranges API
+* Updated service version to `2024-05-04`
 
 ### Breaking Changes
 
 ### Bugs Fixed
 
 ### Other Changes
-* Updated `azblob` version to `1.3.2`
+* Updated `azidentity` version to `1.6.0`
+* Updated `azcore` version to `1.12.0`
 
 ## 1.2.2 (2024-04-09)